@@ -12,8 +12,7 @@ if __name__ == '__main__':
         name='psij-python',
         version=VERSION,
 
-        description='''This is an implementation of the PSI/J (Portable Submission Interface for Jobs)
-        specification.''',
+        description='''This is an implementation of the PSI/J (Portable Submission Interface for Jobs) specification.''',
 
         author='The ExaWorks Team',
         author_email='hategan@mcs.anl.gov',
@@ -43,5 +42,5 @@ if __name__ == '__main__':
         },
 
         install_requires=install_requires,
-        python_requires='>=3.7'
+        python_requires='>=3.8'
     )
