@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Optional, List
 
-from typeguard import check_argument_types
+from typeguard import typechecked
 
 from psij.exceptions import InvalidJobException
 
@@ -45,6 +45,7 @@ class ResourceSpec(ABC):
             raise ValueError()
 
 
+@typechecked
 class ResourceSpecV1(ResourceSpec):
     """This class implements V1 of the PSI/J resource specification."""
 
@@ -53,7 +54,7 @@ class ResourceSpecV1(ResourceSpec):
                  processes_per_node: Optional[int] = None,
                  cpu_cores_per_process: Optional[int] = None,
                  gpu_cores_per_process: Optional[int] = None,
-                 exclusive_node_use: bool = True) -> None:
+                 exclusive_node_use: bool = False) -> None:
         """
         Some of the properties of this class are constrained. Specifically,
         `process_count = node_count * processes_per_node`. Specifying all constrained properties
@@ -81,8 +82,6 @@ class ResourceSpecV1(ResourceSpec):
 
         All constructor parameters are accessible as properties.
         """
-        assert check_argument_types()
-
         self.node_count = node_count
         self.process_count = process_count
         self.processes_per_node = processes_per_node
