@@ -6,7 +6,7 @@ class ExecutorTestParams:
     """A class holding executor, launcher, url pairs."""
 
     def __init__(self, spec: str, queue_name: Optional[str] = None,
-                 project_name: Optional[str] = None,
+                 account: Optional[str] = None,
                  custom_attributes_raw: Optional[Dict[str, Dict[str, object]]] = None) \
             -> None:
         """
@@ -19,8 +19,8 @@ class ExecutorTestParams:
             url are specified, the string should be formatted as "<executor>::<url>".
         queue_name
             An optional queue to submit the job to
-        project_name
-            An optional project name to associate the job with
+        account
+            An optional account to use for billing purposes.
         custom_attributes
         """
         spec_l = re.split(':', spec, maxsplit=2)
@@ -35,7 +35,7 @@ class ExecutorTestParams:
             self.url = None
 
         self.queue_name = queue_name
-        self.project_name = project_name
+        self.account = account
         self.custom_attributes_raw = custom_attributes_raw
         self.custom_attributes: Dict[str, object] = {}
 
